refactor: move objectTreeHeader
